+++ src/main/java/br/com/firstdecision/powercenter/remessa/RegistroSegmentoJ52.java
@@ -5,7 +5,7 @@ public class RegistroSegmentoJ52 {
 	private Integer codigoBanco = 104;
 	private Integer loteServico;
 	private Integer codigoRegistro = 3;
-	private Long nsr;
+	private Integer nsr;
 	private String codigoSegmento = "J";
 	private String usoExclusioFebraban = " ";
 	private Integer codigoMovimenoRemessa;
@@ -43,10 +43,10 @@ public class RegistroSegmentoJ52 {
 	public void setCodigoRegistro(Integer codigoRegistro) {
 		this.codigoRegistro = codigoRegistro;
 	}
-	public Long getNsr() {
+	public Integer getNsr() {
 		return nsr;
 	}
-	public void setNsr(Long nsr) {
+	public void setNsr(Integer nsr) {
 		this.nsr = nsr;
 	}
 	public String getCodigoSegmento() {
